Refactor observation plotting into a modular package
Split the monolithic `apts/observations/plotting.py` into a package structure with specialized mixins for property management, weather plotting, and celestial plotting. This improves code organization and adheres to the Single Responsibility Principle while maintaining backward compatibility.
