+++ apts/observations/plotting/__init__.py
@@ -1,0 +1,7 @@
+from .base import BasePlottingMixIn
+from .weather import WeatherPlottingMixIn
+from .celestial import CelestialPlottingMixIn
+
+
+class PlottingMixIn(BasePlottingMixIn, WeatherPlottingMixIn, CelestialPlottingMixIn):
+    pass
