+++ apts/observations/plotting/celestial.py
@@ -1,49 +1,18 @@
-import logging
 from datetime import datetime
 from typing import TYPE_CHECKING, Optional, cast
 
 if TYPE_CHECKING:
     import matplotlib.figure
+    from ...plotting import NullPlotter, Plotter
 
-    from ..plotting import NullPlotter, Plotter
-
-from ..conditions import Conditions
-from ..constants.plot import CoordinateSystem
-from ..i18n import language_context
-
-if TYPE_CHECKING:
-    from ..observations.base import Observation
-
-logger = logging.getLogger(__name__)
+from ...constants.plot import CoordinateSystem
+from ...i18n import language_context
 
 
-class PlottingMixIn:
+class CelestialPlottingMixIn:
     if TYPE_CHECKING:
-        _plot: Optional["Plotter | NullPlotter"]
-
-    @property
-    def plot(self) -> "Plotter | NullPlotter":
-        if self._plot is None:
-            try:
-                from .. import plotting as apts_plot
-
-                self._plot = apts_plot.Plotter(cast("Observation", self))
-            except ImportError:
-                # Fallback if dependencies are missing or plotting is disabled
-                from ..plotting.dispatcher import NullPlotter
-
-                self._plot = NullPlotter()
-            except Exception as e:
-                # Fallback for any other initialization error
-                logger.warning(f"Failed to initialize plotter: {e}")
-                from ..plotting.dispatcher import NullPlotter
-
-                self._plot = NullPlotter()
-        return self._plot
-
-    @plot.setter
-    def plot(self, value):
-        self._plot = value
+        @property
+        def plot(self) -> "Plotter | NullPlotter": ...
 
     def plot_visible_planets_svg(
         self,
@@ -84,183 +53,6 @@ class PlottingMixIn:
     ):
         with language_context(language):
             return self.plot.planets(dark_mode_override=dark_mode_override, **args)
-
-    def plot_weather(
-        self,
-        dark_mode_override: Optional[bool] = None,
-        language: Optional[str] = None,
-        conditions: Optional[Conditions] = None,
-        **args,
-    ):
-        with language_context(language):
-            return self.plot.weather(
-                dark_mode_override=dark_mode_override, conditions=conditions, **args
-            )
-
-    def plot_clouds(
-        self,
-        dark_mode_override: Optional[bool] = None,
-        language: Optional[str] = None,
-        conditions: Optional[Conditions] = None,
-        **args,
-    ):
-        with language_context(language):
-            return self.plot.clouds(
-                dark_mode_override=dark_mode_override, conditions=conditions, **args
-            )
-
-    def plot_clouds_summary(
-        self,
-        dark_mode_override: Optional[bool] = None,
-        language: Optional[str] = None,
-        **args,
-    ):
-        with language_context(language):
-            return self.plot.clouds_summary(
-                dark_mode_override=dark_mode_override, **args
-            )
-
-    def plot_precipitation(
-        self,
-        dark_mode_override: Optional[bool] = None,
-        language: Optional[str] = None,
-        conditions: Optional[Conditions] = None,
-        **args,
-    ):
-        with language_context(language):
-            return self.plot.precipitation(
-                dark_mode_override=dark_mode_override, conditions=conditions, **args
-            )
-
-    def plot_precipitation_type_summary(
-        self,
-        dark_mode_override: Optional[bool] = None,
-        language: Optional[str] = None,
-        **args,
-    ):
-        with language_context(language):
-            return self.plot.precipitation_type_summary(
-                dark_mode_override=dark_mode_override, **args
-            )
-
-    def plot_temperature(
-        self,
-        dark_mode_override: Optional[bool] = None,
-        language: Optional[str] = None,
-        conditions: Optional[Conditions] = None,
-        **args,
-    ):
-        with language_context(language):
-            return self.plot.temperature(
-                dark_mode_override=dark_mode_override, conditions=conditions, **args
-            )
-
-    def plot_wind(
-        self,
-        dark_mode_override: Optional[bool] = None,
-        language: Optional[str] = None,
-        conditions: Optional[Conditions] = None,
-        **args,
-    ):
-        with language_context(language):
-            return self.plot.wind(
-                dark_mode_override=dark_mode_override, conditions=conditions, **args
-            )
-
-    def plot_pressure_and_ozone(
-        self,
-        dark_mode_override: Optional[bool] = None,
-        language: Optional[str] = None,
-        **args,
-    ):
-        with language_context(language):
-            return self.plot.pressure_and_ozone(
-                dark_mode_override=dark_mode_override, **args
-            )
-
-    def plot_visibility(
-        self,
-        dark_mode_override: Optional[bool] = None,
-        language: Optional[str] = None,
-        conditions: Optional[Conditions] = None,
-        **args,
-    ):
-        with language_context(language):
-            return self.plot.visibility(
-                dark_mode_override=dark_mode_override, conditions=conditions, **args
-            )
-
-    def plot_moon_illumination(
-        self,
-        dark_mode_override: Optional[bool] = None,
-        language: Optional[str] = None,
-        conditions: Optional[Conditions] = None,
-        **args,
-    ):
-        with language_context(language):
-            return self.plot.moon_illumination(
-                dark_mode_override=dark_mode_override, conditions=conditions, **args
-            )
-
-    def plot_fog(
-        self,
-        dark_mode_override: Optional[bool] = None,
-        language: Optional[str] = None,
-        conditions: Optional[Conditions] = None,
-        **args,
-    ):
-        with language_context(language):
-            return self.plot.fog(
-                dark_mode_override=dark_mode_override, conditions=conditions, **args
-            )
-
-    def plot_aurora(
-        self,
-        dark_mode_override: Optional[bool] = None,
-        language: Optional[str] = None,
-        conditions: Optional[Conditions] = None,
-        **args,
-    ):
-        with language_context(language):
-            return self.plot.aurora(
-                dark_mode_override=dark_mode_override, conditions=conditions, **args
-            )
-
-    def plot_seeing(
-        self,
-        dark_mode_override: Optional[bool] = None,
-        language: Optional[str] = None,
-        conditions: Optional[Conditions] = None,
-        **args,
-    ):
-        with language_context(language):
-            return self.plot.seeing(
-                dark_mode_override=dark_mode_override, conditions=conditions, **args
-            )
-
-    def plot_sqm(
-        self,
-        dark_mode_override: Optional[bool] = None,
-        language: Optional[str] = None,
-        conditions: Optional[Conditions] = None,
-        **args,
-    ):
-        with language_context(language):
-            return self.plot.sqm(
-                dark_mode_override=dark_mode_override, conditions=conditions, **args
-            )
-
-    def plot_weather_summary(
-        self,
-        dark_mode_override: Optional[bool] = None,
-        language: Optional[str] = None,
-        conditions: Optional[Conditions] = None,
-        **args,
-    ):
-        with language_context(language):
-            return self.plot.weather_summary(
-                dark_mode_override=dark_mode_override, conditions=conditions, **args
-            )
 
     def plot_sun_and_moon_path(
         self,
